+++ runtime/src/main/java/org/corfudb/runtime/collections/PersistedCorfuTable.java
@@ -6,12 +6,11 @@ import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.corfudb.runtime.object.ICorfuSMR;
 import org.corfudb.runtime.object.ICorfuSMRProxy;
+import org.corfudb.runtime.object.ICorfuSMRProxyMetadata;
 import org.corfudb.runtime.object.ICorfuSMRUpcallTarget;
 
 import javax.annotation.Nonnull;
 import java.util.Map;
-import org.corfudb.runtime.object.ICorfuSMRProxyMetadata;
-
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -36,10 +35,6 @@ public class PersistedCorfuTable<K, V> implements
     }
 
     public PersistedCorfuTable() {}
-
-    public PersistedCorfuTable(ICorfuSMRProxy<DiskBackedCorfuTable<K, V>> proxy) {
-        this.proxy = proxy;
-    }
 
     public PersistedCorfuTable(ICorfuSMRProxy<DiskBackedCorfuTable<K, V>> proxy,
                                ICorfuSMRProxyMetadata proxyMetadata) {
